+++ weblate/trans/models/component.py
@@ -3051,7 +3051,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
 
     def schedule_sync_terminology(self):
         """Trigger terminology sync in the background."""
-        from weblate.glossary.tasks import sync_terminology
+        from weblate.glossary.tasks import sync_glossary_languages, sync_terminology
 
         if self.is_glossary:
             if settings.CELERY_TASK_ALWAYS_EAGER:
@@ -3059,6 +3059,16 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
                 sync_terminology(self.pk, component=self)
             else:
                 transaction.on_commit(lambda: sync_terminology.delay(self.pk))
+
+        else:
+            for glossary in self.project.glossaries:
+                if settings.CELERY_TASK_ALWAYS_EAGER:
+                    # Execute directly to avoid locking issues
+                    sync_glossary_languages(glossary.pk, component=glossary)
+                else:
+                    transaction.on_commit(
+                        lambda: sync_glossary_languages.delay(glossary.pk)
+                    )
 
     def get_unused_enforcements(self):
         from weblate.trans.models import Unit